refactor : InputValidator의 문자열을 처리하는 메소드들을 StringHandler 클래스 안에 독립시킴.
