+++ src/main/java/utils/StringHandler.java
@@ -1,0 +1,37 @@
+package utils;
+
+import java.util.Arrays;
+
+public class StringHandler {
+
+    private StringHandler(){}
+
+    public static boolean isInRange(String input, int min, int max){
+        if(!isNumericString(input)){
+            return false;
+        }
+
+        return Arrays.stream(input.split(""))
+                .map(Integer::parseInt)
+                .allMatch(x-> x >= min && x <= max);
+    }
+
+    public static boolean isValidLength(String input, int len){
+        return input.length() == len;
+    }
+
+    private static boolean isNumericString(String input){
+        try{
+            Integer.parseInt(input);
+        }catch (IllegalArgumentException IAE){
+            return false;
+        }
+        return true;
+    }
+
+    public static boolean isAllDifferent(String input){
+        return Arrays.stream(input.split(""))
+                .distinct()
+                .count() == input.length();
+    }
+}
